requirements.txt
praw~=7.8.1
python-dotenv~=1.0.1
google-generativeai
psycopg2-binary
sqlalchemy~=2.0.36
apscheduler~=3.11.0
protobuf~=5.29.1
requests~=2.32.3
pytest~=8.3.4